// selectores/src/app/countries/interfaces/country.interface.ts
export enum Region {
  Africa = 'Africa',
  Americas = 'Americas',
  Asia = 'Asia',
  Europe = 'Europe',
  Oceania = 'Oceania',
}

export interface SmallCountry {
  name: string;
  cca3: string;
  border: string[];
}

export interface Country {
  name: Name;
  ccn3?: string;
  status: Status;
  idd: Idd;
  capital: string[];
  altSpellings: string[];
  region: string;
  subregion: Subregion;
  languages: Languages;
  latlng: number[];
  landlocked: boolean;
  borders?: string[];
  area: number;
  demonyms: Demonyms;
  cca3: string;
  translations: { [key: string]: Translation };
  flag: string;
  maps: Maps;
  population: number;
  gini?: { [key: string]: number };
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: StartOfWeek;
  capitalInfo: CapitalInfo;
  postalCode: PostalCode;
}

export interface CapitalInfo {
  latlng: number[];
}

export interface Car {
  signs: string[];
  side: Side;
}

export enum Side {
  Left = 'left',
  Right = 'right',
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export interface Currency {
  symbol: string;
  name: string;
}

export interface Demonyms {
  eng: Eng;
  fra: Eng;
}

export interface Eng {
  f: string;
  m: string;
}

export interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

export interface Idd {
  root: string;
  suffixes: string[];
}

export interface Languages {
  ron?: string;
  swe?: string;
  est?: string;
  por?: string;
  slv?: string;
  fra?: string;
  dan?: string;
  ukr?: string;
  deu?: string;
  lit?: string;
  cnr?: string;
  cat?: string;
  ita?: string;
  gsw?: string;
  roh?: string;
  fao?: string;
  ell?: string;
  rus?: string;
  sqi?: string;
  srp?: string;
  hrv?: string;
  ces?: string;
  slk?: string;
  fin?: string;
  ltz?: string;
  eng?: string;
  nrf?: string;
  isl?: string;
  mkd?: string;
  nor?: string;
  nld?: string;
  bel?: string;
  mlt?: string;
  nfr?: string;
  gle?: string;
  lat?: string;
  spa?: string;
  eus?: string;
  glc?: string;
  bul?: string;
  pol?: string;
  hun?: string;
  glv?: string;
  lav?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  tur?: string;
  bos?: string;
}

export interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

export interface Name {
  common: string;
  official: string;
  nativeName: { [key: string]: Translation };
}

export interface Translation {
  official: string;
  common: string;
}

export interface PostalCode {
  format: null | string;
  regex: null | string;
}

export enum StartOfWeek {
  Monday = 'monday',
}

export enum Status {
  OfficiallyAssigned = 'officially-assigned',
  UserAssigned = 'user-assigned',
}

export enum Subregion {
  CentralEurope = 'Central Europe',
  EasternEurope = 'Eastern Europe',
  NorthernEurope = 'Northern Europe',
  SoutheastEurope = 'Southeast Europe',
  SouthernEurope = 'Southern Europe',
  WesternEurope = 'Western Europe',
}